autotest_ui/requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
Django==3.2.8
enum34==1.1.10
execnet==1.9.0
h11==0.12.0
idna==3.3
importlib-metadata==4.8.1
iniconfig==1.1.1
lxml==4.6.3
namedlist==1.8
outcome==1.1.0
packaging==21.0
pluggy==1.0.0
py==1.10.0
pycparser==2.20
pyOpenSSL==21.0.0
pyparsing==2.4.7
pytest==6.2.5
pytest-assume==2.4.3
pytest-dependency==0.5.1
pytest-forked==1.3.0
pytest-ordering==0.6
pytest-rerunfailures==10.2
pytest-xdist==2.4.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==3.10.0.2
urllib3==1.26.7
wsproto==1.0.0
zipp==3.6.0
